move select_types to andi.py
